--- Openshift-EE/pipelines/stages/2-setup/2LP01-openebs-deploy/openebs-deploy
@@ -1,135 +1,0 @@
-#!/bin/bash
-
-pod() {
-sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'cd e2e-openshift && bash Openshift-EE/pipelines/stages/2-setup/2LP01-openebs-deploy/openebs-deploy node '"'$CI_JOB_ID'"'' '"'$CI_PIPELINE_ID'"' '"'$CI_COMMIT_SHA'"' '"'$RELEASE_TAG'"' '"'$RELEASE_BRANCH'"' '"'$E_USER'"' '"'$E_PASSWORD'"' '"'$OPENEBS_MAYA_APISERVER_IMAGE'"' '"'$OPENEBS_PROVISIONER_IMAGE'"' '"'$OPENEBS_SNAPSHOT_CONTROLLER_IMAGE'"' '"'$OPENEBS_SNAPSHOT_PROVISIONER_IMAGE'"' '"'$OPENEBS_IO_JIVA_IMAGE'"' '"'$OPENEBS_MAYA_EXPORTER_IMAGE'"' '"'$OPENEBS_NODE_DISK_MANAGER_IMAGE'"' '"'$OPENEBS_NODE_DISK_OPERATOR_IMAGE'"' '"'$OPENEBS_IO_CSTOR_TARGET_IMAGE'"' '"'$OPENEBS_IO_CSTOR_POOL_IMAGE'"' '"'$OPENEBS_IO_CSTOR_POOL_MGMT_IMAGE'"' '"'$OPENEBS_IO_CSTOR_VOLUME_MGMT_IMAGE'"' '"'$OPENEBS_LOCALPV_PROVISIONER_IMAGE'"' '"'$OPENEBS_ADMISSION_SERVER_IMAGE'"' '"'$OPENEBS_CLEANUP_JOB_IMAGE'"' '"'$DEPLOY_MODE'"' '"'$COMMIT_ID'"' '"'$RC_TAG'"' '"'$HELM_REPO'"' '"'$HELM_BRANCH'"'
-}
-
-node() {
-job_id=$(echo $1)
-pipeline_id=$(echo $2)
-commit_id=$(echo $3)
-releaseTag=$(echo $4)
-releaseBranch=$(echo $5)
-case_id=2LP01
-elastic_user=$(echo $6)
-elastic_password=$(echo $7)
-maya_api_server_image=$(echo $8)
-provisioner_image=$(echo $9)
-snapshot_controller_image=$(echo ${10})
-snapshot_provisioner_image=$(echo ${11})
-jiva_image=$(echo ${12})
-maya_exporter_image=$(echo ${13})
-ndm_image=$(echo ${14})
-ndm_operator_image=$(echo ${15})
-cstor_target_image=$(echo ${16})
-cstor_pool_image=$(echo ${17})
-cstor_pool_mgmt_image=$(echo ${18})
-cstor_volume_mgmt_image=$(echo ${19})
-localpv_provisioner_image=$(echo ${20})
-admission_server_image=$(echo ${21})
-cleanup_job_image=$(echo ${22})
-deploy_mode=$(echo ${23})
-commit_id=$(echo ${24})
-rc_tag=$(echo ${25})
-helm_repo=$(echo ${26})
-helm_branch=$(echo ${27})
-
-time="date"
-current_time=$(eval $time)
-
-present_dir=$(pwd)
-echo $present_dir
-
-echo "releaseTag=$releaseTag"
-
-# Applying e2e-CRD
-echo "***Applying e2e-crd***********"
-kubectl apply -f Openshift-EE/utils/e2e-crd.yml
-
-echo "*****Deploying openebs*******"
-bash Openshift-EE/utils/e2e-cr jobname:openebs-deploy jobphase:Waiting
-bash Openshift-EE/utils/e2e-cr jobname:openebs-deploy jobphase:Running init_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag"
-
-source ~/.profile
-
-echo "Generating test name***************************"
-test_name=$(bash Openshift-EE/utils/generate_test_name testcase=openebs-provision metadata="")
-echo $test_name
-
-## Clone the litmus repo, navigate to litmus root 
-
-git clone https://github.com/openebs/e2e-tests
-cd e2e-tests
-
-export OPENEBS_MAYA_APISERVER_IMAGE=$maya_api_server_image
-export OPENEBS_PROVISIONER_IMAGE=$provisioner_image
-export OPENEBS_SNAPSHOT_CONTROLLER_IMAGE=$snapshot_controller_image
-export OPENEBS_SNAPSHOT_PROVISIONER_IMAGE=$snapshot_provisioner_image
-export OPENEBS_IO_JIVA_IMAGE=$jiva_image
-export OPENEBS_MAYA_EXPORTER_IMAGE=$maya_exporter_image
-export OPENEBS_NODE_DISK_MANAGER_IMAGE=$ndm_image
-export OPENEBS_NODE_DISK_OPERATOR_IMAGE=$ndm_operator_image
-export OPENEBS_IO_CSTOR_TARGET_IMAGE=$cstor_target_image
-export OPENEBS_IO_CSTOR_POOL_IMAGE=$cstor_pool_image
-export OPENEBS_IO_CSTOR_POOL_MGMT_IMAGE=$cstor_pool_mgmt_image
-export OPENEBS_IO_CSTOR_VOLUME_MGMT_IMAGE=$cstor_volume_mgmt_image
-export OPENEBS_LOCALPV_PROVISIONER_IMAGE=$localpv_provisioner_image
-export OPENEBS_ADMISSION_SERVER_IMAGE=$admission_server_image
-export OPENEBS_CLEANUP_JOB_IMAGE=$cleanup_job_image
-export DEPLOY_MODE=$deploy_mode
-export HELM_REPO=$helm_repo
-export HELM_BRANCH=$helm_branch
-export COMMIT_ID=$commit_id
-export RC_TAG=$rc_tag
-
-wget https://raw.githubusercontent.com/openebs/e2e-infrastructure/master/env-update/env.py
-python3 env.py -f providers/openebs/installers/operator/master/litmusbook/run_litmus_test.yml
-
-sed -i -e 's/imagePullPolicy: Always/imagePullPolicy: IfNotPresent/g' \
--e "s/value: '10'/value: '20'/g" \
--e "s/value: master/value: $releaseBranch/g" providers/openebs/installers/operator/master/litmusbook/run_litmus_test.yml
-
-cat providers/openebs/installers/operator/master/litmusbook/run_litmus_test.yml
-
-#################
-## RUNNER MAIN ##
-#################
-
-echo "Applying rbac.yml********************************"
-kubectl apply -f ./hack/rbac.yaml
-kubectl apply -f ./hack/crds.yaml
-
-echo "Running litmus test for openebs deploy.."
-
-run_test=providers/openebs/installers/operator/master/litmusbook/run_litmus_test.yml
-bash ../Openshift-EE/utils/litmus_job_runner label='provider:openebs-runner' job=$run_test
-
-echo "Dumping state of cluster post job run"; echo ""
-bash ../Openshift-EE/utils/dump_cluster_state;
-cd ..
-bash Openshift-EE/utils/event_updater jobname:openebs-deploy $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
-#################
-## GET RESULT  ##
-#################
-rc_val=$(echo $?)
-test_case_id="2LP01"
-
-## Check the test status & result from the litmus result custom resource
-source ~/.profile
-gittoken=$(echo "$github_token")
-testResult=$(kubectl get litmusresult ${test_name} --no-headers -o custom-columns=:spec.testStatus.result)
-
-current_time=$(eval $time)
-bash Openshift-EE/utils/e2e-cr jobname:openebs-deploy jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:$testResult
-
-if [ "$rc_val" != "0" ]; then
-exit 1;
-fi
-
-}
-
-if [ "$1" == "node" ];then
-  node $2 $3 $4 $5 $6 $7 $8 $9 ${10} ${11} ${12} ${13} ${14} ${15} ${16} ${17} ${18} ${19} ${20} ${21} ${22} ${23} ${24} ${25} ${26} ${27} ${28}
-else
-  pod
-fi
