--- Openshift-EE/pipelines/stages/3-functional/3LF02-openebs-localpv-hostpath/openebs-localpv-hostpath
@@ -1,106 +1,0 @@
-#!/bin/bash
-
-
-pod() {
-sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'cd e2e-openshift && bash Openshift-EE/pipelines/stages/3-functional/3LF02-openebs-localpv-hostpath/openebs-localpv-hostpath node '"'$CI_JOB_ID'"'' '"'$CI_PIPELINE_ID'"' '"'$CI_COMMIT_SHA'"' '"'$RELEASE_TAG'"'
-}
-
-node() {
-source ~/.profile
-gittoken=$(echo "$github_token")
-
-job_id=$(echo $1)
-pipeline_id=$(echo $2)
-commit_id=$(echo $3)
-releaseTag=$(echo $4)
-case_id=3LF02
-
-time="date"
-current_time=$(eval $time)
-
-present_dir=$(pwd)
-echo $present_dir
-
-bash Openshift-EE/utils/e2e-cr jobname:openebs-localpv-hostpath jobphase:Waiting
-bash Openshift-EE/utils/e2e-cr jobname:openebs-localpv-hostpath jobphase:Running init_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag"
-
-################
-# LitmusBook 1 #
-################
-
-run_id="localpv";test_name=$(bash Openshift-EE/utils/generate_test_name testcase=busybox-provision-busybox-hostpath metadata=${run_id})
-echo $test_name
-
-cd e2e-tests
-cp apps/busybox/deployers/run_litmus_test.yml run_test_local_pv.yml
-
-sed -i -e 's/app: busybox-litmus/app: deploy-busybox-local-pv/g' \
--e 's/app=busybox-sts/app=local-pv/g' \
--e 's/value: statefulset/value: deployment/g' \
--e 's/value: openebs-cstor-sparse/value: openebs-hostpath/g' \
--e 's/value: openebs-busybox/value: openebs-busybox-hostpath/g' \
--e 's/value: app-busybox-ns/value: busybox-hostpath/g' run_test_local_pv.yml
-
-sed -i '/command:/i \
-          - name: RUN_ID\
-            value: '"$run_id"'\
-' run_test_local_pv.yml
-
-cat run_test_local_pv.yml
-
-bash ../Openshift-EE/utils/litmus_job_runner label='app:deploy-busybox-local-pv' job=run_test_local_pv.yml
-cd ..
-bash Openshift-EE/utils/dump_cluster_state;
-bash Openshift-EE/utils/event_updater jobname:openebs-localpv-hostpath $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
-
-if [ "$?" != "0" ]; then
-exit 1;
-fi
-
-run_id="localpv";test_name=$(bash Openshift-EE/utils/generate_test_name testcase=busybox-deprovision-busybox-hostpath metadata=${run_id})
-echo $test_name
-
-cd e2e-tests
-cp apps/busybox/deployers/run_litmus_test.yml deprovision_local_pv.yml
-
-sed -i -e 's/app: busybox-litmus/app: deprovision-busybox-local-pv/g' \
--e 's/app=busybox-sts/app=local-pv/g' \
--e 's/value: statefulset/value: deployment/g' \
--e 's/value: openebs-cstor-sparse/value: openebs-hostpath/g' \
--e 's/value: openebs-busybox/value: openebs-busybox-hostpath/g' \
--e 's/value: provision/value: deprovision/g' \
--e 's/value: app-busybox-ns/value: busybox-hostpath/g' deprovision_local_pv.yml
-
-sed -i '/command:/i \
-          - name: RUN_ID\
-            value: '"$run_id"'\
-' deprovision_local_pv.yml
-
-cat deprovision_local_pv.yml
-
-bash ../Openshift-EE/utils/litmus_job_runner label='app:deprovision-busybox-local-pv' job=deprovision_local_pv.yml
-cd ..
-bash Openshift-EE/utils/dump_cluster_state;
-bash Openshift-EE/utils/event_updater jobname:openebs-localpv-hostpath $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
-
-rc_val=$(echo $?)
-current_time=$(eval $time)
-
-if [ "$rc_val" != "0" ]; then
-bash Openshift-EE/utils/e2e-cr jobname:openebs-localpv-hostpath jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Fail
-exit 1;
-fi
-
-bash Openshift-EE/utils/e2e-cr jobname:openebs-localpv-hostpath jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Pass
-
-if [ "$rc_val" != "0" ]; then
-exit 1;
-fi
-
-}
-
-if [ "$1" == "node" ];then
-  node $2 $3 $4 $5
-else
-  pod
-fi
