--- Openshift-EE/pipelines/stages/5-cleanup/5LD01-openebs-cleanup/openebs-cleanup
@@ -1,74 +1,0 @@
-#!/bin/bash
-
-pod() {
-sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'cd e2e-openshift && bash Openshift-EE/pipelines/stages/5-cleanup/5LD01-openebs-cleanup/openebs-cleanup node '"'$CI_JOB_ID'"'' '"'$CI_PIPELINE_ID'"' '"'$CI_COMMIT_SHA'"' '"'$RELEASE_TAG'"'
-}
-
-node() {
-job_id=$(echo $1)
-pipeline_id=$(echo $2)
-commit_id=$(echo $3)
-releaseTag=$(echo $4)
-case_id=5LD01
-
-time="date"
-current_time=$(eval $time)
-
-present_dir=$(pwd)
-echo $present_dir
-
-bash Openshift-EE/utils/e2e-cr jobname:openebs-cleanup jobphase:Waiting
-bash Openshift-EE/utils/e2e-cr jobname:openebs-cleanup jobphase:Running init_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag"
-
-test_name=$(bash Openshift-EE/utils/generate_test_name testcase=openebs-deprovision metadata="")
-echo $test_name
-cd e2e-tests
-
-# copy the content of deployer run_litmus_test.yml into a different file to update the test specific parameters.
-cp providers/openebs/installers/operator/master/litmusbook/run_litmus_test.yml cleanup_openebs.yml
-
-#################
-## RUNNER MAIN ##
-#################
-
-
-sed -i -e 's/name: litmus-openebs-runner/name: litmus-openebs-deprovision/g' \
--e 's/provider: openebs-runner/provider: openebs-deprovision/g' \
--e 's/value: provision/value: deprovision/g' cleanup_openebs.yml
-
-cat cleanup_openebs.yml
-
-bash ../Openshift-EE/utils/litmus_job_runner label='provider:openebs-deprovision' job=cleanup_openebs.yml
-bash ../Openshift-EE/utils/dump_cluster_state;
-cd ..
-bash Openshift-EE/utils/event_updater jobname:openebs-cleanup $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
-
-#################
-## GET RESULT  ##
-#################
-rc_val=$(echo $?)
-
-## Check the test status & result from the litmus result custom resource
-source ~/.profile
-gittoken=$(echo "$github_token")
-testResult=$(kubectl get litmusresult ${test_name} --no-headers -o custom-columns=:spec.testStatus.result)
-current_time=$(eval $time)
-
-if [ "$rc_val" != "0" ]; then
-bash Openshift-EE/utils/e2e-cr jobname:openebs-cleanup jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Fail
-exit 1;
-fi
-
-bash Openshift-EE/utils/e2e-cr jobname:openebs-cleanup jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Pass
-
-
-if [ "$rc_val" != "0" ]; then
-exit 1;
-fi
-}
-
-if [ "$1" == "node" ];then
-  node $2 $3 $4 $5
-else
-  pod
-fi
