--- Openshift-EE/utils/e2e-crd.yml
@@ -1,62 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: e2e
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: e2e
-  namespace: e2e
-  labels:
-    name: e2e
----
-# Source: openebs/templates/clusterrole.yaml
-apiVersion: rbac.authorization.k8s.io/v1beta1
-kind: ClusterRole
-metadata:
-  name: e2e
-  labels:
-    name: e2e
-rules:
-- apiGroups: ["*"]
-  resources: ["*"]
-  verbs: ["*"]
----
-apiVersion: rbac.authorization.k8s.io/v1beta1
-kind: ClusterRoleBinding
-metadata:
-  name: e2e
-  labels:
-    name: e2e
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: e2e
-subjects:
-- kind: ServiceAccount
-  name: e2e
-  namespace: e2e
----
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  # name must match the spec fields below, and be in the form: <plural>.<group>
-  name: e2eresults.e2e.io
-spec:
-  # group name to use for REST API: /apis/<group>/<version>
-  group: e2e.io
-  # version name to use for REST API: /apis/<group>/<version>
-  version: v1alpha1
-  # either Namespaced or Cluster
-  scope: Cluster
-  names:
-    # plural name to be used in the URL: /apis/<group>/<version>/<plural>
-    plural: e2eresults
-    # singular name to be used as an alias on the CLI and for display
-    singular: e2eresult
-    # kind is normally the CamelCased singular type. Your resource manifests use this.
-    kind: e2eResult
-    # shortNames allow shorter string to match your resource on the CLI
-    shortNames:
-    - e2er
