--- Openshift-EE/utils/event_updater
@@ -1,23 +1,0 @@
-#!/bin/bash
-set -x
-
-time="date"
-current_time=$(eval $time)
-
-job_name=$(echo $1 | cut -d ":" -f 2)
-test_name=$(echo $2)
-job_id=$(echo $3 | cut -d ":" -f 2)
-pipeline_id=$(echo $4 | cut -d ":" -f 2)
-testcase_id=$(echo $5 | cut -d ":" -f 2)
-
-testResult=$(kubectl get litmusresult ${test_name} --no-headers -o custom-columns=:spec.testStatus.result)
-
-if [[ ${testResult} == "Pass" ]]; then
-echo "      $test_name: Pass" >> result-$job_name.yml
-kubectl apply -f result-$job_name.yml
-echo "***$test_name: Passed******"
-else
-echo "      $test_name: Fail" >> result-$job_name.yml
-bash Openshift-EE/utils/e2e-cr jobname:$job_name jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" test_result:Fail
-kubectl apply -f result-$job_name.yml; exit 1
-fi
