+++ openebs-openshift-e2e/pipelines/stages/3-functional/3LF01-Localpv-selected-device/openebs-localpv-device
@@ -1,0 +1,142 @@
+#!/bin/bash
+
+pod() {
+sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'cd e2e-konvoy && bash openebs-openshift-e2e/pipelines/stages/3-functional/3LF01-openebs-localpv-device/openebs-localpv-device node '"'$CI_JOB_ID'"'' '"'$CI_PIPELINE_ID'"' '"'$CI_COMMIT_SHA'"' '"'$RELEASE_TAG'"'
+}
+
+node() {
+source ~/.profile
+gittoken=$(echo "$github_token")
+
+job_id=$(echo $1)
+pipeline_id=$(echo $2)
+commit_id=$(echo $3)
+releaseTag=$(echo $4)
+case_id=3LF01
+
+time="date"
+current_time=$(eval $time)
+
+present_dir=$(pwd)
+echo $present_dir
+
+oc adm policy add-scc-to-user anyuid system:serviceaccount:busybox-device:default
+
+bash openebs-openshift-e2e/utils/e2e-cr jobname:openebs-localpv-device jobphase:Waiting
+bash openebs-openshift-e2e/utils/e2e-cr jobname:openebs-localpv-device jobphase:Running init_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag"
+
+################
+# e2eBook 1 #
+################
+
+run_id="localpv";test_name=$(bash openebs-openshift-e2e/utils/generate_test_name testcase=busybox-provision-busybox-device metadata=${run_id})
+echo $test_name
+
+cd dynamic-localpv-provisioner/e2e-tests
+cp apps/busybox/deployers/run_e2e_test.yml run_test_local_pv_device.yml
+
+sed -i -e 's/app: busybox-e2e/app: busybox-local-pv-device/g' \
+-e 's/app=busybox-sts/app=local-pv-device/g' \
+-e 's/value: statefulset/value: deployment/g' \
+-e 's/value: openebs-hostpath/value: openebs-device/g' \
+-e 's/value: openebs-busybox/value: openebs-busybox-device/g' \
+-e 's/value: app-busybox-ns/value: busybox-device/g' run_test_local_pv_device.yml
+
+sed -i '/command:/i \
+          - name: RUN_ID\
+            value: '"$run_id"'\
+' run_test_local_pv_device.yml
+
+cat run_test_local_pv_device.yml
+
+bash ../../openebs-openshift-e2e/utils/e2e_job_runner label='app:busybox-local-pv-device' job=run_test_local_pv_device.yml
+cd ../..
+bash openebs-openshift-e2e/utils/dump_cluster_state;
+bash openebs-openshift-e2e/utils/event_updater jobname:openebs-localpv-device $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
+
+if [ "$?" != "0" ]; then
+exit 1;
+fi
+
+################
+# e2eBook 2 #
+################
+
+run_id="test";test_name=$(bash openebs-openshift-e2e/utils/generate_test_name testcase=localpv-selected-device metadata=${run_id})
+echo $test_name
+
+cd dynamic-localpv-provisioner/e2e-tests
+cp experiments/functional/localpv-provisioning-selected-device/run_e2e_test.yml run_local_pv_device.yml
+
+sed -i -e "/name: PVC/{n;s/.*/            value: openebs-busybox-device/g}" \
+-e "/name: APP_NAMESPACE/{n;s/.*/            value: busybox-device/g}" \
+-e "/name: OPERATOR_NS/{n;s/.*/            value: openebs/g}" \
+-e "/name: TEST_CASE_TYPE/{n;s/.*/            value: positive/g}" run_local_pv_device.yml
+
+sed -i '/command:/i \
+          - name: RUN_ID\
+            value: '"$run_id"'\
+' run_local_pv_device.yml
+
+cat run_local_pv_device.yml
+
+bash ../../openebs-openshift-e2e/utils/e2e_job_runner label='app:localpv-selected-device' job=run_local_pv_device.yml
+cd ../..
+bash openebs-openshift-e2e/utils/dump_cluster_state;
+bash openebs-openshift-e2e/utils/event_updater jobname:openebs-localpv-device $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
+
+if [ "$?" != "0" ]; then
+exit 1;
+fi
+
+################
+# e2eBook 3 #
+################
+
+run_id="localpv";test_name=$(bash openebs-openshift-e2e/utils/generate_test_name testcase=busybox-deprovision-busybox-device metadata=${run_id})
+echo $test_name
+
+cd dynamic-localpv-provisioner/e2e-tests
+cp apps/busybox/deployers/run_e2e_test.yml run_test_deprovision_device.yml
+
+sed -i -e 's/app: busybox-e2e/app: deprovision-busybox-local-pv-device/g' \
+-e 's/app=busybox-sts/app=local-pv-device/g' \
+-e 's/value: statefulset/value: deployment/g' \
+-e 's/value: openebs-cstor-sparse/value: openebs-device/g' \
+-e 's/value: openebs-busybox/value: openebs-busybox-device/g' \
+-e 's/value: provision/value: deprovision/g' \
+-e 's/value: app-busybox-ns/value: busybox-device/g' run_test_deprovision_device.yml
+
+sed -i '/command:/i \
+          - name: RUN_ID\
+            value: '"$run_id"'\
+' run_test_deprovision_device.yml
+
+cat run_test_deprovision_device.yml
+
+bash ../../openebs-openshift-e2e/utils/e2e_job_runner label='app:deprovision-busybox-local-pv-device' job=run_test_deprovision_device.yml
+cd ../..
+bash openebs-openshift-e2e/utils/dump_cluster_state;
+bash openebs-openshift-e2e/utils/event_updater jobname:openebs-localpv-device $test_name jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id"
+
+rc_val=$(echo $?)
+current_time=$(eval $time)
+
+if [ "$rc_val" != "0" ]; then
+bash openebs-openshift-e2e/utils/e2e-cr jobname:openebs-localpv-device jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Fail
+exit 1;
+fi
+
+bash openebs-openshift-e2e/utils/e2e-cr jobname:openebs-localpv-device jobphase:Completed end_time:"$current_time" jobid:"$job_id" pipelineid:"$pipeline_id" testcaseid:"$case_id" openebs_version:"$releaseTag" test_result:Pass
+
+if [ "$rc_val" != "0" ]; then
+exit 1;
+fi
+
+}
+
+if [ "$1" == "node" ];then
+  node $2 $3 $4 $5
+else
+  pod
+fi
