+++ openebs-openshift-e2e/pipelines/stages/5-cleanup/5LD02-cluster-cleanup/cluster-cleanup
@@ -1,0 +1,93 @@
+#!/bin/bash
+
+pod() {
+echo $CI_JOB_ID
+###clone e2e-openshift-repo
+echo "cloning e2e-openshift repo*************"
+sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'git clone -b openebs-localpv https://github.com/openebs/e2e-openshift.git'
+sshpass -p $pass ssh -o StrictHostKeyChecking=no $user@$ip -p $port 'cd e2e-openshift && bash Openshift-EE/pipelines/stages/5-cleanup/5LD02-cluster-cleanup/cluster-cleanup node '"'$CI_JOB_ID'"'' '"'$master_ip'"' '"'$worker1_ip'"' '"'$worker2_ip'"' '"'$worker3_ip'"' '"'$worker4_ip'"' '"'$worker5_ip'"' '"'$master_name'"' '"'$worker1_name'"' '"'$worker2_name'"' '"'$worker3_name'"' '"'$worker4_name'"' '"'$worker5_name'"' '"'$master_ip1'"' '"'$master_ip2'"' '"'$master_name1'"' '"'$master_name2'"'
+}
+
+node() {
+job_id=$(echo $1)
+master_ip=$(echo $2)
+master_ip1=$(echo ${14})
+master_ip2=$(ech ${15})
+worker1_ip=$(echo $3)
+worker2_ip=$(echo $4)
+worker3_ip=$(echo $5)
+worker4_ip=$(echo $6)
+worker5_ip=$(echo $7)
+
+master_name=$(echo $8)
+master_name1=$(echo ${16})
+master_name2=$(echo ${17})
+worker1_name=$(echo $9)
+worker2_name=$(echo ${10})
+worker3_name=$(echo ${11})
+worker4_name=$(echo ${12})
+worker5_name=$(echo ${13})
+
+git clone https://github.com/openebs/e2e-tests
+cd e2e-tests
+
+# Replace the VM IP In csv file
+sed -i -e "s/10.12.22.10/$master_ip/g" \
+-e "s/10.12.22.11/$master1_ip/g" \
+-e "s/10.12.22.12/$master2_ip/g" \
+-e "s/10.12.22.13/$worker1_ip/g" k8s/on-prem/openshift-installer/ip.csv
+
+sed -i "/$worker1_ip/a \
+$worker2_ip \
+" k8s/on-prem/openshift-installer/ip.csv
+
+sed -i "/$worker2_ip/a \
+$worker3_ip \
+" k8s/on-prem/openshift-installer/ip.csv
+
+sed -i "/$worker3_ip/a \
+$worker4_ip \
+" k8s/on-prem/openshift-installer/ip.csv
+
+sed -i "/$worker4_ip/a \
+$worker5_ip \
+" k8s/on-prem/openshift-installer/ip.csv
+
+#Replace the Vm names in csv file
+sed -i -e "s/auto1/$master_name/g" \
+-e "s/auto2/$master_name1/g" \
+-e "s/auto3/$master_name2/g" \
+-e "s/auto4/$worker1_name/g" k8s/on-prem/openshift-installer/vm_name.csv
+
+sed -i "/$worker1_name/a \
+$worker2_name \
+" k8s/on-prem/openshift-installer/vm_name.csv
+
+sed -i "/$worker2_name/a \
+$worker3_name \
+" k8s/on-prem/openshift-installer/vm_name.csv
+
+sed -i "/$worker3_name/a \
+$worker4_name \
+" k8s/on-prem/openshift-installer/vm_name.csv
+
+sed -i "/$worker4_name/a \
+$worker5_name \
+" k8s/on-prem/openshift-installer/vm_name.csv
+
+#Replace the snapshot name and exp ip in vars
+sed -i -e 's/snapshot_name: "oc-cluster"/snapshot_name: "cauto"/g' \
+-e 's/esx_ip: "10.12.1.1"/esx_ip: "10.51.1.1"/g' k8s/on-prem/openshift-installer/vars.yml
+
+ansible-playbook k8s/on-prem/openshift-installer/revert_cluster_state.yml 
+
+#Removimg e2e openshift test repo
+cd  && rm -rf e2e-openshift 
+}
+
+if [ "$1" == "node" ];then
+  node $2 $3 $4 $5 $6 $7 $8 $9 ${10} ${11} ${12} ${13} ${14} ${15} ${16} ${17} ${18}
+else
+  pod
+fi
+
