+++ openebs-openshift-e2e/utils/e2e-crd.yml
@@ -1,0 +1,62 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: cie2e
+---
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: cie2e
+  namespace: cie2e
+  labels:
+    name: cie2e
+---
+# Source: openebs/templates/clusterrole.yaml
+apiVersion: rbac.authorization.k8s.io/v1beta1
+kind: ClusterRole
+metadata:
+  name: cie2e
+  labels:
+    name: cie2e
+rules:
+- apiGroups: ["*"]
+  resources: ["*"]
+  verbs: ["*"]
+---
+apiVersion: rbac.authorization.k8s.io/v1beta1
+kind: ClusterRoleBinding
+metadata:
+  name: cie2e
+  labels:
+    name: cie2e
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: cie2e
+subjects:
+- kind: ServiceAccount
+  name: cie2e
+  namespace: cie2e
+---
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  # name must match the spec fields below, and be in the form: <plural>.<group>
+  name: cie2eresults.e2e.io
+spec:
+  # group name to use for REST API: /apis/<group>/<version>
+  group: e2e.io
+  # version name to use for REST API: /apis/<group>/<version>
+  version: v1alpha1
+  # either Namespaced or Cluster
+  scope: Cluster
+  names:
+    # plural name to be used in the URL: /apis/<group>/<version>/<plural>
+    plural: cie2eresults
+    # singular name to be used as an alias on the CLI and for display
+    singular: cie2eresult
+    # kind is normally the CamelCased singular type. Your resource manifests use this.
+    kind: cie2eResult
+    # shortNames allow shorter string to match your resource on the CLI
+    shortNames:
+    - cie2er
