+++ openebs-openshift-e2e/utils/e2e-result.yml
@@ -1,0 +1,17 @@
+---
+apiVersion: e2e.io/v1alpha1
+kind: cie2eResult
+metadata:
+  # name of the e2e testcase
+  name: <test-name>
+spec:
+  testResult: <test-result>
+  testcaseId: <case-id>
+  testStatus:
+    phase: <test-phase>
+    start_time: <test-init>
+    end_time: <test-finish>
+    build_id: <openebs-id>
+    job_id: <job-id>
+    pipeline_id: <pipeline-id>
+    event:
