+++ openebs-openshift-e2e/utils/get_e2e_result
@@ -1,0 +1,25 @@
+#!/bin/bash
+set -x 
+
+e2e_test_name=$1
+
+time="date"
+current_time=$(eval $time)
+
+## Check the test status & result from the e2e result custom resource
+
+testStatus=$(kubectl get e2eresult ${e2e_test_name} --no-headers -o custom-columns=:spec.testStatus.phase) && \
+testResult=$(kubectl get e2eresult ${e2e_test_name} --no-headers -o custom-columns=:spec.testStatus.result); retcode=$?
+bash openebs-openshift-e2e/utils/error_handler ${retcode} msg="Unable to find e2e result custom resource, exiting" action="exit"
+
+if [[ ${testStatus} == "completed" ]]; then
+  if [[ ${testResult} == "Pass" ]]; then
+    bash openebs-openshift-e2e/utils/e2e-cr jobname:$2 jobphase:Completed end_time:"$current_time" test_result:Pass
+    echo "TEST: PASS"
+  else
+    bash openebs-openshift-e2e/utils/e2e-cr jobname:$2 jobphase:Completed end_time:"$current_time" test_result:Fail
+    echo "TEST: FAILED"; exit 1
+  fi
+else
+  echo "Test Execution was aborted"; exit 1
+fi
